refactor: change user to account
