+++ Trinity.Application/Services/AccountsService.cs
@@ -3,7 +3,6 @@ using SecureIdentity.Password;
 using System.Threading.Tasks;
 using Trinity.Application.Contracts;
 using Trinity.Application.DTOs.Accounts;
-using Trinity.Application.DTOs.Users;
 using Trinity.Application.Exceptions.Accounts;
 using Trinity.Application.Wrappers;
 using Trinity.Domain.Entities.Accounts;
@@ -13,21 +12,21 @@ namespace Trinity.Application.Services
 {
     public class AccountsService : IAccountsService
     {
-        private readonly IStaticPersistence<Accounts> usersStaticPersistence;
-        private readonly IBasePersistence<Accounts> usersBasePersistence;
+        private readonly IStaticPersistence<Accounts> accountsStaticPersistence;
+        private readonly IBasePersistence<Accounts> accountsBasePersistence;
         private readonly IPasswordHasherWrapper passwordHasher;
         private readonly ITokenService tokenService;
         private readonly IMapper mapper;
 
         public AccountsService(
-            IStaticPersistence<Accounts> usersStaticPersistence,
-            IBasePersistence<Accounts> usersBasePersistence,
+            IStaticPersistence<Accounts> accountsStaticPersistence,
+            IBasePersistence<Accounts> accountsBasePersistence,
             IPasswordHasherWrapper passwordHasher,
             ITokenService tokenService,
             IMapper mapper)
         {
-            this.usersStaticPersistence = usersStaticPersistence;
-            this.usersBasePersistence = usersBasePersistence;
+            this.accountsStaticPersistence = accountsStaticPersistence;
+            this.accountsBasePersistence = accountsBasePersistence;
             this.passwordHasher = passwordHasher;
             this.tokenService = tokenService;
             this.mapper = mapper;
@@ -35,7 +34,7 @@ namespace Trinity.Application.Services
 
         public async Task<AccountsOutput> SignUpAsync(AccountsSignUpInput accountInput)
         {
-            Accounts? accountExists = await this.usersStaticPersistence.GetByEmailAsync(accountInput.Email);
+            Accounts? accountExists = await this.accountsStaticPersistence.GetByEmailAsync(accountInput.Email);
 
             if (accountExists != null)
             {
@@ -47,13 +46,13 @@ namespace Trinity.Application.Services
 
             AccountsOutput accountOutput = this.mapper.Map<AccountsOutput>(account);
 
-            await this.usersBasePersistence.AddAsync(account);
+            await this.accountsBasePersistence.AddAsync(account);
             return accountOutput;
         }
 
         public async Task<TokenOutput> SignInAsync(AccountsSignInInput accountInput)
         {
-            Accounts? account = await this.usersStaticPersistence.GetByEmailAsync(accountInput.Email) ?? throw new AccountsException("Email not registered.");
+            Accounts? account = await this.accountsStaticPersistence.GetByEmailAsync(accountInput.Email) ?? throw new AccountsException("Email not registered.");
 
             if (!this.passwordHasher.Verify(account.PasswordHash, accountInput.Password))
             {
